+++ fabric/src/main/java/club/iananderson/seasonhud/fabric/event/ClientEvents.java
@@ -16,7 +16,7 @@ public class ClientEvents {
   private static void registerKeyInputs() {
     ClientTickEvents.END_CLIENT_TICK.register(client -> {
       if (KeyBindings.seasonhudOptionsKeyMapping.consumeClick()) {
-        MainConfigScreen.open();
+        MainConfigScreen.getInstance().open();
       }
     });
   }