+++ forge/src/main/java/club/iananderson/seasonhud/forge/event/ClientEvents.java
@@ -22,7 +22,7 @@ public class ClientEvents {
     @SubscribeEvent
     public static void onKeyInput(InputEvent.Key Event) {
       if (KeyBindings.seasonhudOptionsKeyMapping.consumeClick()) {
-        MainConfigScreen.open();
+        MainConfigScreen.getInstance().open();
       }
     }
   }